refactor(auth): absorb api-key package into @reloop/auth (expand) (#46)
Move generation/hashing helpers to dependency-light @reloop/auth/apikey
and validation to @reloop/auth/apikey/validate for the upcoming middleware
plugin. Leave @reloop/apikey as a thin re-export shim so existing
importers keep working; deletion lands in the contract ticket.

Adds unit tests for generate/hash/key-start helpers and isolation
coverage that the helpers graph excludes Redis/DB/Elysia. Characterization
tripwire still passes via the shim.
