+++ packages/auth/package.json
@@ -6,6 +6,8 @@
 		"./client": "./src/client.ts",
 		"./server": "./src/server.ts",
 		"./types": "./src/types.ts",
+		"./apikey": "./src/apikey/index.ts",
+		"./apikey/validate": "./src/apikey/validate.ts",
 		"./roles": "./src/roles.ts",
 		"./permissions": "./src/permissions.ts",
 		"./platform-permissions": "./src/platform-permissions.ts"
