+++ packages/auth/src/apikey/index.ts
@@ -1,0 +1,12 @@
+/**
+ * `@reloop/auth/apikey` — dependency-light generation/hashing helpers only.
+ * Validation (Redis/DB) lives at `@reloop/auth/apikey/validate`.
+ */
+export {
+	API_KEY_LENGTH,
+	API_KEY_PREFIX,
+	generateApiKey,
+	getApiKeyCacheKey,
+	getKeyStart,
+	hashApiKey,
+} from "./helpers";
